게시글 삭제 기능, 게시글 상태 수정 기능의 권한 변경, Eager 로딩 설정

+++ src/boards/boards.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { Board } from './board.entity';
 import { BoardStatus } from "./board-status.enum";
 import { InjectRepository } from '@nestjs/typeorm';
@@ -6,6 +6,7 @@ import { Repository } from 'typeorm';
 import { UpdateBoardDto } from './dto/update-board.dto';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { User } from 'src/auth/user.entity';
+import { UserRole } from 'src/auth/user-role.enum';
 
 @Injectable()
 export class BoardsService {
@@ -60,16 +61,27 @@ export class BoardsService {
     }
 
     // 특정 번호의 게시글 삭제
-    async deleteBoardById(id: number): Promise<void> {
+    async deleteBoardById(id: number, user: User): Promise<void> {
         const foundBoard = await this.getBoardById(id); // 게시글 조회
+        // 작성자와 요청한 사용자가 같은지 확인
+        if (foundBoard.user.id !== user.id) {
+            throw new UnauthorizedException(`You do not have permission to delete this board`);
+        }
         await this.boardsRepository.remove(foundBoard); // 게시글 삭제
     }
 
-    // 특정 번호의 게시글의 일부 수정
-    async updateBoardStatusById(id: number, status: BoardStatus): Promise<void> {
-        const result = await this.boardsRepository.update(id, { status });
-        if (result.affected === 0) {
-          throw new NotFoundException(`Board with ID ${id} not found`);
+    // 특정 번호의 게시글의 일부 수정(관리자가 부적절한 글을 비공개로 설정)
+    async updateBoardStatusById(id: number, status: BoardStatus, user: User): Promise<void> {
+        // 관리자인지 확인
+        if (user.role === UserRole.ADMIN) {
+            // 관리자는 상태를 변경할 수 있음
+            const result = await this.boardsRepository.update(id, { status });
+            if (result.affected === 0) {
+                throw new NotFoundException(`There's no updated record or Board with ID ${id} not found`);
+            }
+        } else {
+            // 일반 사용자는 상태 변경 권한이 없음
+            throw new UnauthorizedException(`You do not have permission to update the status of this board`);
         }
     }
